chore: Moved db/files.py into Document Manager
