+++ db/database_documents.py
@@ -1,4 +1,10 @@
 import psycopg
+from pydantic import BaseModel
+
+class Document(BaseModel):
+    id:str
+    owner:int
+    filename:str
 
 
 class DatabaseDocumentManager:
@@ -28,9 +34,28 @@ class DatabaseDocumentManager:
             await self.__conn.commit()
             return doc_id
 
-
-    async def delete_document(self, id: int):
-        delete_query = "DELETE FROM document_data.documents WHERE ID = %s"
+    async def get_all_files_for_user(self,user_id: int) -> list[Document]:
+        sel_query="SELECT * FROM document_data.documents WHERE owner = %s"
         async with self.__conn.cursor() as cur:
-            await cur.execute(delete_query, (id,))
-            await self.__conn.commit()
+            await cur.execute(sel_query,(user_id,),)
+            files: list[Document] = []
+            results = await cur.fetchall()
+            for file in results:
+                files.append(
+                    Document(
+                        id=file[0],
+                        owner=file[1],
+                        filename=file[2],
+                    )
+                )
+        return files
+
+    async def delete_file_by_id(self, user_id:int, document_id: str):
+        del_query = "DELETE FROM document_data.documents WHERE id=%s AND document_id=%s"
+        async with self.__conn.cursor() as cur:
+            await cur.execute(
+                del_query,
+                (user_id,
+                document_id,),
+            )
+        await self.__conn.commit()
